Module kirk::crew

Traits that abstract how to send jobs to workers.

Modules

channel
A crew that uses std::sync::mpsc.

deque
A crew that uses the lock-free Chase-Lev deque implemented in crossbeam.

Traits

Crew
A collection of workers that can be given jobs.

Parameters
Options used to alter the behavior or configuration of a crew.

Worker
Workers perform jobs in separate threads as part of a crew.