Module kirk::crew::deque

A crew that uses the lock-free Chase-Lev deque implemented in crossbeam.

Each worker runs this loop in a separate thread:

Try to steal a job

If successful, perform it, and become "hot"

If shutting down, break

If lost a race or no jobs, "cool down"

Possibly yield or sleep

When a hot worker "cools down," it becomes "warm", with a retry count of zero; an already-warm worker increases the retry count. Eventually, the retry threshold may be exceeded, and the worker becomes "cold".

The temperature determines the action at step five: a hot worker immediately continues the loop; a warm one cooperatively yields to another thread; and cold ones sleep. The goal is to allow workers to progress unhindered as long as there are jobs, but reduce excessive CPU usage during periods when there are little to none.

A cold worker may also become hot again if it loses a race to steal a job, since this strongly indicates that another job is ready, but remains cold as long as it finds the queue empty.

Structs

Deque
A Crew that uses a lock-free Chase-Lev deque.

DequeWorker
The Crew Member for Deque.

Options
Parameters to adjust the size and behavior of the crew.