Struct kirk::crew::deque::Options

pub struct Options {
    pub retry_threshold: u32,
    pub cold_interval: Duration,
    pub num_workers: usize,
}

Parameters to adjust the size and behavior of the crew.

The current defaults for retry threshold and cold interval--32 and 1ms--were chosen arbitrarily. Experimentation may be prudent.

Fields

retry_threshold

How many times may a worker fail to acquire a job before it becomes "cold" and sleeps for cold_interval between subsequent attempts.

cold_interval

The minimum length of time a worker will sleep when it is cold.

num_workers

The number of workers to create in the pool.