Trait kirk::crew::Crew

pub trait Crew {
    type Job: Job;
    type Member: Worker;
    type Settings: Parameters;
    fn new(settings: Self::Settings) -> Self;
    fn hire(&mut self) -> Self::Member;
    fn give<F>(&mut self, job: F) where Self::Job: From<F>;
    fn stop(&mut self);
}

A collection of workers that can be given jobs.

Associated Types

type Job: Job

type Member: Worker

type Settings: Parameters

Required Methods

fn new(settings: Self::Settings) -> Self

Construct a crew from its parameters.

fn hire(&mut self) -> Self::Member

Construct a new worker.

fn give<F>(&mut self, job: F) where Self::Job: From<F>

Send a job to the workers.

fn stop(&mut self)

Tell the workers to stop after completing all outstanding jobs.