Trait kirk::crew::Worker

pub trait Worker: Send {
    fn run(&mut self);
}

Workers perform jobs in separate threads as part of a crew.

Pool moves each Worker into its own thread, which is why they must be Send.

Required Methods

fn run(&mut self)

Execute the worker's main loop.

Typecially, this will look somewhat like:

Possibly acquire a message.

If successful, perform it.

If error or message is 'stop', break.

Each crew implements these differently, but all fulfill this contract.