Struct kirk::Pool

pub struct Pool<C> where C: Crew {
    // some fields omitted
}

A crew of workers that perform jobs on separate threads.

Even after dropped, the workers continue processing outstanding jobs, then halt when there are no more. Any shared resources of the crew would get dropped then.

If scoped, the enclosing thread will only block once the crossbeam::scope ends, since it must join on all threads it has spawned for safety.

Methods

impl<'scope, C, J> Pool<C> where J: Job + 'scope, C: Crew<Job=J> + 'scope

fn scoped(scope: &Scope<'scope>, settings: C::Settings) -> Pool<C>

Create a new, scoped pool.

fn push<F>(&mut self, f: F) where J: From<F>

Give a job to the pool's crew.

impl<C, J> Pool<C> where J: Job + 'static, C: Crew<Job=J> + 'static

fn new(settings: C::Settings) -> Pool<C>

Create an new, unscoped pool.

When dropped, each worker thread is joined, to ensure that all outstanding jobs