Struct kirk::Task

pub struct Task<'a>(_);

A boxed closure that can be performed as a job.

This allows pools to execute any kind of job, but with the increased cost of dynamic invocation.

Warning

Because of the epoch-based memory management of crossbeam, it's possible for Task pools to deadlock if the closures own data that cause another thread to block. For example, this would hang if the line drop(tx); were removed from the tasks' closure.

use std::sync::mpsc::channel;
use std::thread;
use kirk::{Pool, Task, Deque};
use kirk::crew::deque::Options;

let rx = {
    let (tx, rx) = channel();
    let options = Options::default();
    let mut pool = Pool::<Deque<Task>>::new(options);
    for i in 0..10 {
        let tx = tx.clone();
        pool.push(move || {
           tx.send(i + 1).unwrap();
           drop(tx);
        });
    }
    rx
};
thread::spawn(move || {
    for _ in rx.iter() {}
});